Change GameObject X,Y to tile ones

+++ AgribattleArena/AgribattleArena.Engine/ForExternalUse/Synchronization/ObjectInterfaces/IActiveDecoration.cs
@@ -14,8 +14,8 @@ namespace AgribattleArena.Engine.ForExternalUse.Synchronization.ObjectInterfaces
         float Health { get; }
         int? OwnerId { get; }
         bool IsAlive { get; }
-        float X { get; }
-        float Y { get; }
+        int X { get; }
+        int Y { get; }
         float Z { get; }
         float MaxHealth { get; }
         List<TagSynergy> Armor { get; }
