+++ AgribattleArena/AgribattleArena.Engine/Synchronizers/SynchronizationObjects/ActiveDecoration.cs
@@ -18,8 +18,8 @@ namespace AgribattleArena.Engine.Synchronizers.SynchronizationObjects
         public List<TagSynergy> Armor { get; }
         public int? OwnerId { get; }
         public bool IsAlive { get; }
-        public float X { get; }
-        public float Y { get; }
+        public int X { get; }
+        public int Y { get; }
         public float Z { get; }
 
         public ActiveDecoration (Objects.ActiveDecoration decoration)
